Moved payments tier types to types folder
